+++ crc32trim.h
@@ -32,7 +32,7 @@ extern "C" {
 extern unsigned long crc32_trim_trailing(
     unsigned long crc1,
     unsigned long crc2,
-    int len2
+    long len2
 );
 
 #ifdef __cplusplus
